SmartWeb-Assistant/requirements.txt
streamlit
requests
beautifulsoup4
html2text
torch
langchain
langchain-community
langchain-huggingface
langchain-groq
python-dotenv
qdrant-client
